Updated schemas and parsers for unsigned values

+++ catapult-sdk/src/model/ModelSchemaBuilder.js
@@ -43,8 +43,8 @@ class ModelSchemaBuilder {
 			// region block
 			blockHeader: {
 				size: ModelType.int,
-				version: ModelType.int,
-				network: ModelType.int,
+				version: ModelType.uint8,
+				network: ModelType.uint8,
 				type: ModelType.int,
 				height: ModelType.uint64,
 				timestamp: ModelType.uint64,
@@ -57,7 +57,7 @@ class ModelSchemaBuilder {
 				receiptsHash: ModelType.binary,
 				stateHash: ModelType.binary,
 				beneficiaryAddress: ModelType.binary,
-				feeMultiplier: ModelType.int
+				feeMultiplier: ModelType.uint32
 			},
 			blockHeaderMetadata: {
 				hash: ModelType.binary,
@@ -84,20 +84,20 @@ class ModelSchemaBuilder {
 			finalizedBlock: {
 				height: ModelType.uint64,
 				hash: ModelType.binary,
-				finalizationEpoch: ModelType.uint,
-				finalizationPoint: ModelType.uint
+				finalizationEpoch: ModelType.uint32,
+				finalizationPoint: ModelType.uint32
 			},
 
 			finalizationProof: {
-				version: ModelType.uint,
-				finalizationEpoch: ModelType.uint,
-				finalizationPoint: ModelType.uint,
+				version: ModelType.uint32,
+				finalizationEpoch: ModelType.uint32,
+				finalizationPoint: ModelType.uint32,
 				height: ModelType.uint64,
 				hash: ModelType.binary,
 				messageGroups: { type: ModelType.array, schemaName: 'messageGroup' }
 			},
 			messageGroup: {
-				stage: ModelType.uint,
+				stage: ModelType.uint32,
 				height: ModelType.uint64,
 				hashes: { type: ModelType.array, schemaName: ModelType.binary },
 				signatures: { type: ModelType.array, schemaName: 'bmTreeSignature' }
@@ -118,8 +118,8 @@ class ModelSchemaBuilder {
 
 			transaction: {
 				size: ModelType.int,
-				version: ModelType.int,
-				network: ModelType.int,
+				version: ModelType.uint8,
+				network: ModelType.uint8,
 				type: ModelType.int,
 				deadline: ModelType.uint64,
 				maxFee: ModelType.uint64
@@ -167,7 +167,7 @@ class ModelSchemaBuilder {
 				addressHeight: ModelType.uint64,
 				publicKey: ModelType.binary,
 				publicKeyHeight: ModelType.uint64,
-				accountType: ModelType.int,
+				accountType: ModelType.uint8,
 				supplementalPublicKeys: { type: ModelType.object, schemaName: 'supplementalPublicKey' },
 				importance: ModelType.uint64,
 				importanceHeight: ModelType.uint64,
@@ -183,7 +183,7 @@ class ModelSchemaBuilder {
 			activityBucket: {
 				startHeight: ModelType.uint64,
 				totalFeesPaid: ModelType.uint64,
-				beneficiaryCount: ModelType.int,
+				beneficiaryCount: ModelType.uint32,
 				rawScore: ModelType.uint64
 			},
 			mosaic: {
@@ -198,8 +198,8 @@ class ModelSchemaBuilder {
 			},
 			votingPublicKey: {
 				publicKey: ModelType.binary,
-				startEpoch: ModelType.uint,
-				endEpoch: ModelType.uint
+				startEpoch: ModelType.uint32,
+				endEpoch: ModelType.uint32
 			},
 
 			// endregion
@@ -220,7 +220,7 @@ class ModelSchemaBuilder {
 				db: ModelType.string
 			},
 			nodeInfo: {
-				version: ModelType.int,
+				version: ModelType.uint8,
 				roles: ModelType.int,
 				port: ModelType.int,
 				networkIdentifier: ModelType.int,
